move blog form into its own component
